% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rosenbrock.R
\name{do_iterations}
\alias{do_iterations}
\title{A function for performing the iterative procedure and printing the results}
\usage{
do_iterations(num_iter, x1_init, x2_init, tol)
}
\arguments{
\item{num_iter}{integer, must be greater than 0}

\item{x1_init}{scalar}

\item{x2_init}{scalar}

\item{tol}{positive real}
}
\value{
numeric vector containing current x1 iterate, current x2 iterate,
value of objective at current iterate, number of iterations performed, and
a list of all values of the objective function, up to and including the
current value.
}
\description{
A function for performing the iterative procedure and printing the results
}
\examples{
do_iterations(1,1,1,0.5)
}
