% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rosenbrock.R
\name{f}
\alias{f}
\title{The original Rosenbrock objective function which we wish to majorize}
\usage{
f(x1, x2)
}
\arguments{
\item{x1}{scalar}

\item{x2}{scalar}
}
\value{
scalar output
}
\description{
The original Rosenbrock objective function which we wish to majorize
}
\examples{
f(1,1)
}
